requirements.txt
cudatoolkit=10.2.89
cudnn=8.2.1.32
numpy=1.21.4
python=3.7.11
pytorch=1.10.0
tensorboard=2.7.0
torchvision=0.11.1
absl-py=0.15.0
grpcio=1.42.0
markdown=3.3.6
pillow=9.0.1
protobuf=3.19.1
pyyaml=6.0
six=1.16.0
tqdm=4.62.3
werkzeug=2.0.3
